Add demo of MapML for BC Albers "BCTILE" (temporary only - unsure how stable the definition of BCTILE might be).
Change name of M.MapML to M.MapMLFeatures, because it was a bit confusing
to explain, inasmuchas there is a M.MapMLLayer which is its 'parent' layer/object.

Add extent management code that allows to include a mapml file from the
same domain as the html etc resources, but which doesn't contain an
extent element.  The approach taken here is not correct an will be 
replaced (soon - working on allowing MapML content inside the <layer-></layer->
element so that authors can *easily* add placemarks etc.)